a very simple pipe reader & writer
a purpose of this thing is to have a bit more flexible pipe that should not be closed in case errors occurred. This is right what I need - keep pipe until client disconnects, but also being able to pass errors like io.EOF (otherwise it's a real crap). Also this shit solves buffers allocating (we pass a slice that refers to some already allocated array, really good for memory - no extra buffers that actually takes really lots of memory)