improvements
tiny improvements. For example, `Readable()` method now takes not pointer to instance but a copy of instance. Also added big and detailed description of what is this pipe is, and how to use it. Now in Read() we do not defer decrement, but do it instantly, as it anyway cannot panic (even if it will it's exceptional case, so we don't care about it)