additions
this is a file that contains responses renderer. Not all the potential of this thing is not revealed but in future it will be. It is supposed to be like a "client session" containing some useful variables for some features in newer protocols (or even for HTTP/1, for example, avoiding some behaviours for older versions)