improvements
since now content encodings and default headers are handled by application instead of router. Content codings, for example, are now completely inaccessible for router because it does not needs it, also default headers are constructed in application. IMHO default headers and content codings are better to be set from application, not from router