fixes
now router also doesn't know about renderer, it is initialised and used primarily by application and http server. This is made because router is single for all the users, renderer appeared to be also - but it is not supposed to be single for all the clients. So decided to move renderer somewhere higher to initialise it per-user