improvements
simplified headers a bit: now value parameters are not parsing, this is headers job now. But header values are now appending not by char, but by sequence - this means that instead of appending one character at once, waiting until the sequence ends, and appending all the sequence. This much optimizes the process, first of all for long header values (20+ bytes)