updates & feature: added ChunkedTE attribute
added `ChunkedTE` attribute to the request object that marks whether request body is chunked or not. Also now in case content length is 0, we are not waiting for a nil byteslice from channel. This removes a need in `FinalizeBody()` hack in parser and optimises requests without body processing for 260ns (due to benchmarks on my local machine)