updates: moved to another package, and other code improvements
moved request and response objects to http package, because `types` package is just a legacy from my very old webserver. Also request object now carries request context and response object, context for decreasing handlers' signature size, and response for better memory management. Because of old design, it was impossible to avoid allocating a hashmap on every response with user headers because response object was a global object. But now it is built-in the request object that allows us manually clear it. Moreover, clearing response headers now is very cheap - because response headers are now just a flat strings array so rendering headers is very simple and fast now