refactored so no more channels
no more channels are used for streaming body processing. Now everything is done in a single execution flow by splitting the parsing process into 2 independent parts - request line and headers, and request body. To avoid situations when we've read more than we should so this affects another part of parsing - we can unread it using built-in client-object mechanism