now using an array of methods instead of a map
previously router used to store a two-dimensional map, where the first layer stands for endpoints, and the second one for a set of methods accepted by the endpoint. Now instead of a secod-dimensional map using just a static-sized array, which is being accessed by index represented by the method enum. This doesn't affect results as I supposed to, but still makes some difference