renamed router.Fabric to router.Builder and Fabric.Initialize() to Builder.Build()
the reason is simple - Fabric looks java'ish, every time I see it I want to vomit. The same with Initialize() - it just doesn't look right. Now we have Builder, which is logical and straightly conveys its point (in case of inbuilt it's indeed a builder), and whose method Build() looks just so much clean and correct.